refactor: store the session executor in the registry, not the endpoint
