+++ crates/burn-remote/src/client/mod.rs
@@ -2,6 +2,7 @@ mod base;
 mod channel;
 mod custom_op;
 mod runner;
+pub(crate) mod runtime;
 pub(crate) mod service;
 
 pub use base::*;
